Match cclevcontrol_update (98.7->100%)
Cast the three getObjGroupStatus() vtable-result null checks to (u8):
the callee returns a byte and the target masks it (clrlwi r0,r3,24;
cmplwi) before comparing, not a signed cmpwi (recipes #11/#124).

+++ src/main/dll/CC/dll_018B_cclevcontrol.c
@@ -210,17 +210,17 @@ void cclevcontrol_update(int obj)
         SCGameBitLatch_UpdateInverted((SCGameBitLatchState*)(state + 1), 0x80, -1, -1, 0x24, 0xea);
     }
     if (GameBit_Get(0x3d6) != 0
-        && (*gMapEventInterface)->getObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1f) != 0)
+        && (u8)(*gMapEventInterface)->getObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1f) != 0)
     {
         (*gMapEventInterface)->setObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1f, 0);
     }
     if (GameBit_Get(0x161) != 0
-        && (*gMapEventInterface)->getObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1e) == 0)
+        && (u8)(*gMapEventInterface)->getObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1e) == 0)
     {
         (*gMapEventInterface)->setObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1e, 1);
     }
     if (GameBit_Get(0x3d7) != 0
-        && (*gMapEventInterface)->getObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1d) == 0)
+        && (u8)(*gMapEventInterface)->getObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1d) == 0)
     {
         (*gMapEventInterface)->setObjGroupStatus(((GameObject*)obj)->anim.mapEventSlot, 0x1d, 1);
     }
